+++ yash-cli/CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - External dependency versions:
     - Rust 1.75.0 → 1.77.0
+- Internal dependency versions:
+    - yash-syntax 0.9.0 → 0.10.0
 - The shell now shows the prompt before reading the input in the interactive mode.
   To achieve this, the `startup::prepare_input` function now applies the
   `yash_prompt::Prompter` decorator to the returned source input.